feat(RHOAIENG-78901): Show fast and unsupported labels in deployed models table (#8810)
* feat(RHOAIENG-78901): consolidate version labels into DeploymentResourceVersionLabels

Consolidate ServingRuntimeVersionLabel into DeploymentResourceVersionLabels so
that fast-version and limited-support labels appear everywhere the runtime
version label appears, including the deployed models table.

- Inline the version Label directly in DeploymentResourceVersionLabels and
  delete ServingRuntimeVersionLabel
- Add getVersion prop (defaults to getServingRuntimeVersion) for extensibility
- Add versionStatus, templateStatus, and scopeLabel slots so callers can render
  additional labels inside the same LabelGroup (required because PF LabelGroup
  does not space children correctly when they are nested in fragments)
- Set numLabels={10} to prevent PF's default 3-label collapse behavior
- Replace all ServingRuntimeVersionLabel usages with DeploymentResourceVersionLabels
- Add unit test for custom getVersion prop


* feat(RHOAIENG-78901): show fast and unsupported labels in deployed models table

Replace DeploymentResourceVersionLabels component with a
renderDeploymentResourceVersionLabels function that returns an array of Label
elements. Callers spread the result into their own LabelGroup as direct
children, working around a PF bug where Labels nested inside fragments lose
their spacing in LabelGroup.

- Convert DeploymentResourceVersionLabels component to
  renderDeploymentResourceVersionLabels function returning keyed Label elements
- Add getVersion option (defaults to getServingRuntimeVersion) for extensibility
- Replace ServingRuntimeVersionLabel usage in KServe and llmd-serving with the
  new function, so fast-version and limited-support labels appear in the
  deployed models table
- Deprecate ServingRuntimeVersionLabel (retained for legacy ServingRuntimeModal)
- Set numLabels={5} on LabelGroups to prevent PF's default 3-label overflow
- Leave ServingRuntimeTemplateSection unchanged (deprecated code path)


* fix: use array expression instead of spread for version labels in LabelGroup

Remove the {...spread} syntax when rendering version labels in LabelGroup
children. React flattens keyed arrays identically to spread elements, and
the simpler {array} syntax is more conventional.


* fix: remove spread syntax and add numLabels to all version label LabelGroups

Remove {...spread} syntax in favor of {array} expressions — React flattens
keyed arrays identically and the simpler syntax is more conventional.

Add numLabels={5} to all LabelGroups containing version labels to prevent
PF's default 3-label overflow collapse.


---------
